% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClusteringRCode_v5g.R
\name{AIC.eClust}
\alias{AIC.eClust}
\title{AIC for edge clustering}
\usage{
\method{AIC}{eClust}(A, eCl)
}
\arguments{
\item{A}{igraph object}

\item{eCl}{object returned from 'eClustGEM'}
}
\description{
Compute the AIC for edge clustering object
}
