% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClusteringRCode_v5g.R
\name{AICc.eClust}
\alias{AICc.eClust}
\title{AICc for edge clustering}
\usage{
AICc.eClust(A, eCl)
}
\arguments{
\item{A}{igraph object}

\item{eCl}{object returned from 'eClustGEM'}
}
\description{
Compute the AICc for edge clustering object
}
