% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClusteringRCode_v5g.R
\name{BIC.eClust}
\alias{BIC.eClust}
\title{BIC for edge clustering}
\usage{
\method{BIC}{eClust}(A, eCl)
}
\arguments{
\item{A}{igraph object}

\item{eCl}{object returned from 'eClustGEM'}
}
\description{
Compute the BIC for edge clustering object
}
