% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClusteringRCode_v5g.R
\name{eClustGEM}
\alias{eClustGEM}
\title{Edge clustering}
\usage{
eClustGEM(A, K = 2, p = 2, maxIter = 1000, eps = 1e-04,
  CGSteps = 25, QNSteps = 25, WApprox = ecount(A) > 1000,
  nSubsamp = max(vcount(A), 500), UVApprox = FALSE,
  UVEstMethod = c("CG", "QN"), BKNInit = FALSE, randUVInit = TRUE,
  a_u = 1, b_u = 1, a_s = 1, b_s = 1, a_v = a_u, b_v = b_u,
  a_r = a_s, b_r = b_s, a_0 = 1)
}
\arguments{
\item{A}{igraph object}

\item{K}{integer.  Number of clusters}

\item{p}{integer. Dimension of latent space}

\item{maxIter}{integer. Maximum number of iterations of the GEM algorithm}

\item{eps}{numeric. Convergence threshold for relative change of log posterior}

\item{CGSteps}{integer.  Maximum number of conjugate gradient steps during each iteration}

\item{QNSteps}{integer.  Maximum number of quasi-Newton steps during each iteration}

\item{WApprox}{logical. Should subsampling approximation be used in updating W?}

\item{UVApprox}{logical. Should subsampling approximation be used in updating U and V?}

\item{BKNInit}{should initialization be done via Ball Karrer, and Newman (2011)?}

\item{randUVInit}{Should the initialization of U and V be random?}

\item{a_u, }{b_u, a_s, b_s, a_v, b_v, a_r, b_r, a_0 numeric.  Hyperparameters.}

\item{UVEstMethod}{character, either 'CG' for conjugate gradient (default) or 'QN' for quasi-Newton}
}
\description{
'eClustGEM' implements the generalized EM algorithm for clustering edges of a network using a latent space approach
}
