\name{LSEC-package}
\alias{LSEC-package}
\alias{LSEC}
\docType{package}
\title{
\packageTitle{LSEC}
}
\description{
\packageDescription{LSEC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LSEC}
\packageIndices{LSEC}
What goes here
}
\author{
\packageAuthor{LSEC}

Maintainer: \packageMaintainer{LSEC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
